Ensure characters activate/deactive appropriately
We haven't before set characters as active or not when doing login and
logout in the server program.  We set characters as deactivated when
loading them from the database, but didn't consider it any other time.

This has other implications:  we've seen segfaults with objects in
motion after the owners logged out.  Part of this was due to not being
able to stop movement or rotation when we wanted to, which is no
longer a problem.  However, if a body is still in motion after a user
logs out, the motion should stop very quickly and the object should be
removed from the motion queues.  The motion pool worker needs more
tweaking to eliminate unnecessary work in the case that a user logged
out while the object was waiting in the queue, but we're most of the
way there.

Now as users login and take control of an object, we activate it, and
when they logout, we deactivate it.  We don't yet tell clients to get
rid of a deactivated object, but we'll get there.