Bail out of motion pool if object is not moving
We haven't bothered testing if an object is even moving before we
start making assumptions that it is, and doing sector changes and
whatnot.  If, say, an object has been deactivated while it was waiting
in the motion queue, we would "move" it and send updates, even though
there would have been nothing to do.

To simplify, we'll only do movement if the object is still moving.  We
still check at the end, whether to reinsert into the motion queue, but
in context of having already moved a moving object.