Add a deleted-object packet type
We need to be able to inform users when objects should be removed.  It
could be because the objects are destroyed, or users logout.  The
object position update would be the most reasonable place for it, but
there doesn't seem to be a good set of parameters which would be
impossible to attain in normal operation, that could function as a
flag to require object removal.

We'll use the base of the position update packet - type, version,
sequence, object ID - with a separate packet type ID.  Not being able
to reuse anything is a little sad, but we're still working on basic
functionality, so no harm done.

We won't opt to bump the packet version number.  We're still very much
before anybody would use this, and we're only adding, not altering.