Move generation of update packets into game object
The update pool worker has been generating the update packets to now,
but since it can't really find out the status of a given object as far
as being active or not, it might actually make more sense for the
thing that *does* know about itself, the game object, to do the packet
generation.  That way, if we're deactivated, we can generate a new
object deletion packet, and otherwise we'll do a normal position
update.

This won't handle the situation in which we delete an object which
will *not* be staying around, for instance something which isn't
associated with a character and will simply be deleted.  Not sure how
that will need to look.