Add a define for supported protocol version
Instead of having the magic constant everywhere, let's use an actual
defined constant.  That way, if we change things, we at least won't
have to search and replace everywhere.

When we do rev protocol version numbers, it's likely that we'd have to
do a bunch of stuff anyway, since we write all the packets mostly by
hand anyway.  But at least this will prevent a little bit of churn.