Convert to use boost::locale for l10n
GNU Gettext is a C library, and doesn't fit very well into the C++
idiom.  Fortunately the Boost libraries have locale handling, which
fits into the idiom very well, is fully compatible with Gettext, and
also doesn't need to link with Gettext.

Doing sensing of whether it's available in the configure file is less
than straightforward, because the standard AC_SEARCH_LIBS is very
C-centric.  Fortunately it's not difficult to sense, just more verbose
than it otherwise might be.  The code snippet that we use is taken
almost verbatim from the boost::locale documentation.

Because it doesn't make sense to maybe have locale handling or maybe
not, we'll make it always on.  So now boost::locale is a hard
requirement to be able to build the client or utility programs.