Start up locale handling properly in client programs
When we made the big conversion to use boost::locale, we neglected to
do the same conversion on the setup in the r9client and r9_keygen
programs.  This is mostly a cookbook change from the boost::locale
documentation.

We also found that the locale directory was completely unset in
makefiles, so the client programs weren't getting the right parameters
either.  Fortunately that's a simple change.