Clarify wordings of error messages
We had a mix of "couldn't do <whatever>" and "error doing <whatever>",
which, well, is not consistent.  We'll standardize on "error doing".

There were also a handful of other messages which we could streamline
and simplify, and even a couple from a couple different files which we
could make the same.