Convert strerror calls to system_error where possible
We discovered that we're using the GNU version of strerror_r, which
returns the string that we should use.  That return value is not
guaranteed to be the same string that we made for the error message.
On our development platform, it isn't, so we get a whole lot of
nothing in our error messages.

C++11 adds the std::system_error, which can mostly replace the
"convert an errno-type value into a string" functionality, and
conveniently combines it with an exception.  Fewer calls = better.

Not all of our uses of strerror_r can be converted in this way, so
we'll fix those uses in a future commit.