Repair misuse of strerror_r
We're using the GNU version of strerror_r, which returns a char
pointer to the appropriate string to use.  That string is *not*
guaranteed to be the same string buffer that we pass in, and on our
development platform it indeed was not, so our log messages had
nothing where there was supposed to be an error string.

We'll keep passing in the buffer, because you never know, but we'll
start using the return value of strerror_r as we should be.  And as
expected, it is now working.