Add tests to CodeQL control file
We might as well run the unit tests, since we're switching our builds
to CodeQL.  We can turn on profiling and debugging as well, for good
measure; maybe that will help the analysis (which seems to do nothing
at the moment).

In case we get a failure, we'll make sure to cat out the log for that
failure.