Update for more modern build environments
Newer `python3-config` data sets understand `--embed`, which has
slightly different flags than the plain `--ldflags`, and if present,
will not build properly without them.  We'll try the one, and if the
newer flag is unrecognized, we'll try the old style.

Doing tests of whether an external CuddlyGL library worked was falling
victim to missing libraries, specifically the Pthread library.  We
move the check for it earlier, and include the lib in the `LIBS` set
for our configure test.