Improve client i18n strings
We can annotate the placeholders within strings to be translated, to
help translators figure out just what will be in those inserted
sections.

We had been missing a few files which had translatable strings, so
those are now being included.  None of the headers which were listed
in the files-to-be-translated had translatable strings, so we dropped
them.