Move build to local action definition
We want to restructure our build/test actions a bit, to shrink the
number of actions we're doing, and also to potentially shorten the
time to do them.

We've been doing CodeQL on *every* push, which seems overkill, and
also takes much longer.  We just want that to happen on pull requests
into main.  The rest of the time, a plain build/test is sufficient.

We move the actual build and test to an action file, and `uses` it
everywhere we need.  The new build-and-test workflow happens on all
pushes, and the codeql workflow now only happens on PRs into main.