Remove last updated from comment headers
Having the last updated date/time stamps in every file causes more
trouble than it's worth.  Reverts are more painful, merges are more
painful, and stashing/restoring is more painful.  It made sense back
before we were using version control, but now git keeps track of all
that stuff for us.

We also update the copyright years, turning them into a range of years
that the file was changed.  We're not counting this commit, because
nothing of interest is actually changing.