Handle octree subtree creation exceptions
The octree constructor can throw an exception, if the r/w lock can not
be initialized.  When we create new subtrees, we could run into this
situation, and instead of just failing out, we should try to keep
running.  If we end up with a lot of objects in an octant, performance
may suffer, but at least we'll still be running.

Handling such blowups in the zone is a whole other kettle of fish.  We
don't allocate any octrees to start unless there are objects in them,
so we could get into a situation where we have objects that have moved
into a sector (represented by an individual octree), and we can't make
a structure to contain them.  This will have to be a separate commit
or set of commits, because it's likely to be complex.  Honestly, at
that point our environment may be compromised enough that it actually
makes sense to crash?